Make sure the Ramses+AHF handler does DM cross-matching only, to avoid failure on attempt to bridging AMR cells
